requirements.txt
requests
beautifulsoup4
lxml
jsonschema
llama-parse
python-dotenv
wikipedia
openai
pydantic
faiss-cpu
numpy
sickle
